requirements.txt
ipykernel
python-dotenv
openai
langchain
langchain-community
langchain-openai
langchain-google-genai
google-generativeai
langgraph
google-adk
crewai
nest_asyncio
gymnasium
stable_baselines3
pygame
markdown
python-docx
